move src/text_to_speec.ts to src/lib/
